requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-google-genai==0.0.6
ollama==0.1.6
chromadb==0.4.22
pydantic==2.5.3
python-dotenv==1.0.0
PyPDF2==3.0.1